fix: apply professional ES module compatibility fixes for error handling
- Reverted test utilities back to CommonJS (module.exports) following Jest conventions
- Inlined error detection logic in setupTests.mjs to avoid cross-module system imports
- Removed unnecessary error-detection.js file to eliminate build conflicts
- Fixed ESLint no-require-imports error while maintaining full functionality
- Ensures robust error handling without compromising ES module compatibility

This resolves the build error while following professional Jest/ES module practices
and maintains all error detection capabilities for production safety.
